+++ cli/CHANGELOG.md
@@ -14,6 +14,7 @@ _Released 9/25/2024_
 
 **Misc:**
 
+- Cypress now consumes [geckodriver](https://firefox-source-docs.mozilla.org/testing/geckodriver/index.html) to help automate the Firefox browser instead of [marionette-client](https://github.com/cypress-io/marionette-client). Addresses [#30217](https://github.com/cypress-io/cypress/issues/30217). Addressed in [#30250](https://github.com/cypress-io/cypress/pull/30250).
 - Pass along the related log to the `createSnapshot` function for protocol usage. Addressed in [#30244](https://github.com/cypress-io/cypress/pull/30244).
 
 **Dependency Updates:**